(feat) create subcomponent file for filter buttons and put OperatorButton and Initial button in there
